niffler_e_2_e_tests_python/requirements.txt
selene>=2.0.0rc9
pytest
requests
pkce
faker
allure-pytest
python-dotenv
sqlmodel
pydantic